==> Rmd/01-mapping.Rmd <==
---
title: "Skin tape striping - mapping"
author: "Tu Hu"
date: "12/13/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
pacman::p_load("dplyr", "stringr", "purrr", "tidyr", "ggplot2")
```

enabled by ultra-affordable bulk RNA barcoding and sequencing (BRB-seq).

- reduced bioinfomratics resources

- reduced cost

## Mapping

```{r fastq files}
sample <-
  openxlsx::read.xlsx("../data/stp_sorting.xlsx") %>% 
  as_tibble()

fn <- 
  list.files("../data/fastq/", pattern = ".fastq.gz") %>% 
  str_remove("_R(1|2)_001.fastq.gz") %>% 
  unique()
writeLines(fn, "../_tmp/fn.txt")

batch <-
  fn %>% str_sub(1,1)
writeLines(batch, "../_tmp/batch.txt")  
```

We received in total `r length(fn)` * 2 (paired-end) files.

### Batch information
```{r}
batch_info <- 
  openxlsx::read.xlsx("../data/batch_info.xlsx") %>% 
  as_tibble() %>% 
  filter(PCR.batch != "-") %>% 
  group_by(PCR.batch) %>% 
  select(Name = sample_id, 
         B1 = UMI.primer.sequence) %>% 
  tidyr::nest() %>% 
  mutate(fn = paste0("../_tmp/batch_info_", PCR.batch %>% str_extract("\\d{1}"), ".txt"))

purrr::map2(batch_info$data,
            batch_info$fn, ~ readr::write_tsv(.x, .y))
```


### QC

```{bash fastqc}
module load parallel/20200522 jdk/17 perl/5.30.2 fastqc/0.11.9
cd ..
ID_PATH=_tmp/fn.txt
cat ${ID_PATH} | parallel "fastqc data/fastq/{}_R1_001.fastq.gz data/fastq/{}_R2_001.fastq.gz -o data/fastqc"

```

```{bash multiqc}
module load anaconda3/4.4.0
multiqc ../data/fastqc -o ../data/fastqc
```

```{r average sequencing depth}
total_read <- 
  readr::read_tsv("../data/fastqc/multiqc_data/multiqc_fastqc.txt") %>% 
  select(Sample, `Total Sequences`) %>% 
  mutate(Sample = Sample %>% str_remove("_R(1|2)_001")) %>% 
  distinct() %>% 
  summarise(sum_read = sum(`Total Sequences`)) %>% 
  pull(sum_read)

total_read / nrow(sample)
```

### Construct reference genome

First, download the latest reference genome from Ensembl.

```{bash download ref genome from ensemble}
mkdir -p ../data/reference
wget -P ../data/reference ftp://ftp.ensembl.org/pub/release-105/fasta/homo_sapiens/dna/Homo_sapiens.GRCh38.dna.primary_assembly.fa.gz
gzip -d ../data/reference/Homo_sapiens.GRCh38.dna.primary_assembly.fa.gz
wget -P ../data/reference ftp://ftp.ensembl.org/pub/release-105/gtf/homo_sapiens/Homo_sapiens.GRCh38.105.gtf.gz
gzip -d ../data/reference/Homo_sapiens.GRCh38.105.gtf.gz
```

```{bash}
module load jre/1.8.0 brbseqtools/1.6.0
```

```{bash constract ref genome}
module load star/2.7.9a
mkdir -p ../data/STAR_Index/
cd ../data
STAR --runMode genomeGenerate --genomeDir STAR_Index/ --genomeFastaFiles reference/Homo_sapiens.GRCh38.dna.primary_assembly.fa --sjdbGTFfile reference/Homo_sapiens.GRCh38.105.gtf --runThreadN 30
```

I first merge four lanes into one fastq file.

```{bash merge fastq}
cd /home/projects/ku_00015/people/tuhu/multiomics-ad-sts/data/fastq
mkdir -p ../fastq_merged
for i in {1..5}
do
   echo forward ${i}
   cat ${i}_S${i}*_R1_001.fastq.gz > ../fastq_merged/${i}_R1.fastq.gz
   
   echo reverse ${i}
   cat ${i}_S${i}*_R2_001.fastq.gz > ../fastq_merged/${i}_R2.fastq.gz
done
```

```{bash STAR alignment reverse read}
module load gcc/9.4.0 star/2.7.9a 
cd /home/projects/ku_00015/people/tuhu/multiomics-ad-sts/data
mkdir -p BAM_merged
for i in {1..5}; do STAR --genomeDir STAR_Index/ \
--runMode alignReads \
--outFilterMultimapNmax 1 \
--readFilesCommand zcat \
--outSAMtype BAM Unsorted \
--outFileNamePrefix BAM_merged/${i} \
--readFilesIn fastq_merged/${i}_R2.fastq.gz \
--runThreadN 20 ; done
```

```{bash use forward read to map samples}
cd /home/projects/ku_00015/people/tuhu/multiomics-ad-sts/data
mkdir -p brbmatrix
module load jre/1.8.0 brbseqtools/1.6.0 
for i in {1..5}; do brbseqtools CreateDGEMatrix \
-f fastq_merged/${i}_R1.fastq.gz \
-b BAM_merged/${i}Aligned.out.bam \
-c ../_tmp/batch_info_${i}.txt \
-gtf reference/Homo_sapiens.GRCh38.105.gtf \
-p B???????????????????? \
-o brbmatrix/${i} ; done
```
