==> Rmd/04-TapeStripvsBiopsy.Rmd <==
---
title: "04-Compare with full-thickness biopsy"
author: "Tu Hu"
date: "2/21/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
pacman::p_load("tidybulk", "tidySummarizedExperiment",
               "tidyr", "dplyr", "stringr",
               "limma", "SummarizedExperiment", 
               "DESeq2", "BiocParallel",
               "tibble")
```

## load data
```{r}
se_biopsy <-
  readRDS("../../multiomics-ad-phd/data/se.rds")

se_tapestriping <-
  readRDS("../data/se_qc.rds")
```


We compared tape striping data with punch biopsy data that is sampled from the same skin site.
The first obvious difference is that, in deep-sequenced full-thickness biopsy, we were able to detect expression of more than 15000 genes, within which 2000 genes were detected as differentially expressed between conditions. But, from tape striping BRB-seq, we were only able to detect 2310 genes, by aggregating all genes, we were only able to detect 

## Comparison
```{r}
se_biopsy_summarise <-
  se_biopsy %>% as_tibble() %>% 
  group_by(subject, visit, skin_type, feature) %>% 
  summarise(biopsy_expr = mean(counts_scaled))


```
