==> Rmd/05-AD_signature.Rmd <==
---
title: "05-AD_signature"
author: "Tu Hu"
date: "2/24/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
pacman::p_load("dplyr", "tidyr", "purrr", "tidybulk", "tidySummarizedExperiment",
               "ComplexHeatmap", "ggplot2")
```

## AD signature
```{r load benchmark data}
benchmark <- 
  readRDS("../data/benchmark_table.rds") %>% 
  group_by(contrast) %>% nest %>% 
  mutate(gene_up = map(data, ~ .x$`Gene (overexpressed)` %>% unlist %>% unique %>% discard(~.x == "")),
         gene_down = map(data, ~ .x$`Gene (downexpressed)` %>% unlist %>% unique %>% discard(~.x == "")))


se_qc <- readRDS("../data/se_qc.rds")
se_qc_34 <- se_qc[,se_qc$qc_0_4 > 2]
se_qc_cpm <- se_qc_34 %>% edgeR::cpmByGroup(group = se_qc_34$skin_type, log = T)

gene_up <- benchmark$gene_up %>% unlist %>% unique
gene_up_plot <- 
  purrr::discard((se_qc_cpm[gene_up,3] > se_qc_cpm[gene_up,1]) & (se_qc_cpm[gene_up,3] > se_qc_cpm[gene_up,2]), ~.x == FALSE) %>% names

gene_down <- benchmark$gene_down %>% unlist %>% unique
gene_down_plot <- 
  purrr::discard((se_qc_cpm[gene_down,3] < se_qc_cpm[gene_down,1]) & (se_qc_cpm[gene_down,3] < se_qc_cpm[gene_down,2]), ~.x == FALSE) %>% names

library("gplots")

heatmap_DEG <- 
  se_qc_cpm[c(gene_up_plot, gene_down_plot),] %>% 
  t() %>% scale %>% 
  t %>% 
  heatmap.2(col = bluered(100), 
            density.info = "none", trace = "none", 
            dendrogram = "row", keysize = 1, 
            margins = c(5,5), lhei = c(.8,7))

heatmap_DEG <- 
  se_qc_cpm[c(gene_up_plot, gene_down_plot),] %>% 
  t() %>% scale %>% 
  t %>% 
  heatmap.2() 


jpeg("../figure/heatmap_DEG.jpeg", width = 20, 
     height = 30, units = "cm", res = 300)
  se_qc_cpm[c(gene_up_plot, gene_down_plot),] %>% 
  t() %>% scale %>% 
  t %>% 
  heatmap.2(col = bluered(100), density.info = "none", trace = "none", dendrogram = "row", keysize = 1, margins = c(5,5), lhei = c(.8,7))
dev.off()

# gene_description <- 
#   gprofiler2::gconvert(c(gene_up_plot, gene_down_plot)) %>% 
#   as_tibble() %>% select(name, description)
# openxlsx::write.xlsx(gene_description, "../data/gene_description.xlsx")
```

```{r heatmap}
se_qc_scale <- se_qc[,se_qc$qc_0_4>2] %>% scale_abundance()
colData(se_qc_scale)$biopsy_area <- 
  colData(se_qc_scale)$biopsy_area %>% 
  str_replace_all("_", " ")

se_qc_deg <- 
  se_qc_scale[c(gene_up_plot, gene_down_plot), ]
se_qc_deg$skin_type <- 
  forcats::fct_relevel(se_qc_deg$skin_type,
                       c("HC", "NL", "LS"))


heatmap_data_m_top100 <-
  scale(se_qc_deg %>% assay(2) %>% t() %>% log1p(),
        center = TRUE, scale = TRUE) %>% t()

sample_site <- 
  colData(se_qc_deg) %>% as_tibble() %>% pull(biopsy_area) %>% unique()
sample_site_color <- RColorBrewer::brewer.pal(length(sample_site), "Pastel2")
names(sample_site_color) <- sample_site

heatmap_annotation <- 
  HeatmapAnnotation(
    `Tissue state` = colData(se_qc_deg) %>% as_tibble() %>% pull(skin_type), 
    `Anatomic region` = colData(se_qc_deg) %>% as_tibble() %>% pull(biopsy_area),
    col = list(`Tissue state` = c("HC" = "#91cf60",
                                   "NL" = "#eb8b9b", 
                                   "LS" = "#eb2d0c"),
               `Anatomic region` = sample_site_color),
    annotation_legend_param = 
      list(`Tissue state` = list(nrow = 1, title_gp = gpar(fontsize = 12),
                                labels_gp = gpar(fontsize = 10)),
           `Anatomic region` = list(nrow = 1, title_gp = gpar(fontsize = 12),
                                    labels_gp = gpar(fontsize = 10))
                                   ))

heatmap_top100 <- ComplexHeatmap::Heatmap(heatmap_data_m_top100,
                                          name = "gene expression",
                                          column_dend_reorder = TRUE,
                        # clustering_method_columns = "complete",
                        show_column_names = FALSE,
                        show_row_names = TRUE,
                        show_row_dend = FALSE,
                        row_names_gp = gpar(fontsize = 9),
                        top_annotation = heatmap_annotation,
                        heatmap_legend_param = list(
                          title_gp = gpar(fontsize = 12),
                          labels_gp = gpar(fontsize = 10)
                          ),
                        column_split = factor(se_qc_deg$skin_type, levels = c("LS", "NL", "HC")),
                        cluster_column_slices = F,
                        row_title_rot = 0,
                        row_title_side = "right")

jpeg(filename = "../figure/heatmap_DEG_all.jpeg", width = 720*4, 
     height = 624*4, quality = 100, res = 300)
draw(heatmap_top100, heatmap_legend_side = "right",
     annotation_legend_side = "top", merge_legend = FALSE)
dev.off()
```

```{r}
se_qc_up %>% 
  as_tibble() %>% 
  group_by(skin_type, feature) %>% 
  summarise(expr = mean(counts_scaled)) %>% 
  tidyHeatmap::heatmap(feature, skin_type, expr, 
                       palette_value = c("blue", "white", "red"))
```
