==> Rmd/06-BiopsyvsTape.Rmd <==
---
title: "06-comparision_BiopsyvsTape"
author: "Tu Hu"
date: "2/25/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
pacman::p_load("tidybulk", "tidySummarizedExperiment",
               "tidyr", "dplyr", "stringr",
               "limma", "SummarizedExperiment", 
               "DESeq2", "BiocParallel",
               "tibble", "ggpubr", "purrr", "forcats")
core_n <- future::availableCores()
register(MulticoreParam(ifelse(core_n <= 8, core_n - 2, core_n / 2)))
```

## Load data
```{r}
se_tape <- readRDS("../data/se_qc.rds")
se_tape <- se_tape[, se_tape$qc_0_4 > 2]

se_biopsy <- readRDS("../../multiomics-ad-phd/data/se.rds")

g_to_compare <-
  rownames(se_tape)[rownames(se_tape) %in% rownames(se_biopsy)]

g_to_compare_FC <-
  readRDS("../data/benchmark_table.rds") %>% group_by(contrast) %>% nest %>% 
  mutate(gene_up = map(data, ~ .x$`Gene (overexpressed)` %>% unlist %>% unique %>% discard(~.x == "")),
         gene_down = map(data, ~ .x$`Gene (downexpressed)` %>% unlist %>% unique %>% discard(~.x == ""))) %>% 
  ungroup %>% 
  select(gene_up, gene_down) %>% unlist %>% unique
```

## skin tape data
```{r skin tape data matrix}
colData(se_tape)$subject_visit <- 
  str_c(colData(se_tape)$subject_id,
        colData(se_tape)$visit, sep = "_")
colData(se_tape)$subject_visit_skintype <- 
  str_c(colData(se_tape)$subject_id,
        colData(se_tape)$visit,
        colData(se_tape)$skin_type, sep = "_")
se_tape_t <- 
  se_tape[g_to_compare, ] %>% scale_abundance()
expr_tape_m <-
  assay(se_tape_t, 2)
colnames(expr_tape_m) <-
  colData(se_tape)$subject_visit_skintype
```

```{r skin tape data - FC}
se_tape_AD <- se_tape_t[, se_tape_t$skin_type %in% c("LS", "NL")]

se_tape_AD_FC <- 
  se_tape_AD[g_to_compare_FC, 
             se_tape_AD$subject_visit %in% 
               (se_tape_AD$subject_visit %>% table %>% 
                  discard(~ .x == 1) %>% names)] %>% 
  as_tibble() %>% 
  mutate(counts_scaled = ifelse(counts_scaled == 0, .5, counts_scaled)) %>% 
  arrange(skin_type) %>% 
  group_by(subject_visit, feature) %>% nest %>% 
  mutate(FC_tape = map_dbl(data, ~ log2(.x$counts_scaled[2] / .x$counts_scaled[1]))) %>% 
  select(-data)


```

## biopsy data
```{r collapse libraries, warning=FALSE}
se_biopsy_g_to_comapre <- se_biopsy[g_to_compare,]

colData(se_biopsy_g_to_comapre)$subject_visit_skintype <- 
  str_c(colData(se_biopsy_g_to_comapre)$subject,
        colData(se_biopsy_g_to_comapre)$visit,
        colData(se_biopsy_g_to_comapre)$skin_type, sep = "_")

colData(se_biopsy_g_to_comapre)$subject_visit <- 
  str_c(colData(se_biopsy_g_to_comapre)$subject,
        colData(se_biopsy_g_to_comapre)$visit, sep = "_")

sp_to_collapse <- 
  colData(se_biopsy_g_to_comapre)$subject_visit_skintype %>% unique

expr_biopsy <- 
  bplapply(1:length(sp_to_collapse),
       function(sp){
  index <- 
    which(colData(se_biopsy_g_to_comapre)$subject_visit_skintype ==
            sp_to_collapse[sp])
  expr <- assay(se_biopsy_g_to_comapre, 2)[,index]
  expr_mean <- if (length(index) > 1) {
    rowMeans(expr)
  } else {
    expr
  }
  return(expr_mean)
})
names(expr_biopsy) <- sp_to_collapse
expr_biopsy_m <- 
  expr_biopsy %>% purrr::reduce(bind_cols) %>% as.matrix()
rownames(expr_biopsy_m) <- g_to_compare
colnames(expr_biopsy_m) <- sp_to_collapse
```

```{r skin biopsy FC}
se_biopsy_AD_FC <-
  expr_biopsy_m %>% as_tibble(rownames = "gene_name") %>% 
  pivot_longer(cols = !gene_name, names_to = "subject_visit_skintype",
               values_to = "counts") %>% 
  mutate(counts = counts + .5) %>% 
  filter(gene_name %in% g_to_compare_FC) %>% 
  mutate(subject_visit = subject_visit_skintype %>% 
           str_extract("(AD|CO)_\\d{2}_0\\d{1}"),
         skin_type = subject_visit_skintype %>% 
           str_extract("(LS|NL|HC)")) %>% 
  arrange(skin_type) %>% 
  group_by(subject_visit, gene_name) %>% nest %>% 
  mutate(nrow = map_dbl(data, ~ .x %>% nrow)) %>% 
  filter(nrow == 2) %>% 
  mutate(FC = map_dbl(data, ~ log2(.x$counts[1] / .x$counts[2]) )) %>% 
  select(feature = gene_name, subject_visit, FC_biopsy = FC)
```

### FC
```{r warning=FALSE}
FC_s_correlation <- 
  se_tape_AD_FC %>% ungroup() %>% 
  left_join(se_biopsy_AD_FC, by = c("feature", "subject_visit")) %>% 
  filter(!is.na(FC_biopsy)) %>% group_by(subject_visit) %>% nest() %>% 
  mutate(
    cor_spearman = map(data, ~ cor.test(.x$FC_tape, .x$FC_biopsy, 
                                        method = "spearman") %>% 
                         broom::tidy())) %>% 
  select(subject_visit, cor_spearman) %>% 
  unnest
```

```{r warning=F}
FC_g_correlation <- 
  se_tape_AD_FC %>% ungroup() %>% 
  left_join(se_biopsy_AD_FC, by = c("feature", "subject_visit")) %>% 
  filter(!is.na(FC_biopsy)) %>% group_by(feature) %>% nest() %>% 
  mutate(cor_spearman = map(data, ~ cor.test(.x$FC_tape, .x$FC_biopsy, 
                                             method = "spearman") %>% broom::tidy())) %>% 
  select(feature, cor_spearman) %>% unnest %>% 
  arrange(-estimate)
```


## Correlation coefficient 

### Sample

```{r sample-wise correlation}
tape_biopsy_s_res <- 
  bplapply(which(colnames(expr_tape_m) %in% intersect(
  colnames(expr_tape_m), colnames(expr_biopsy_m))), 
  function(sp_index){
  sp_name <- colnames(expr_tape_m)[sp_index]
  tape_s <- expr_tape_m[,sp_name]
  biopsy_s <- expr_biopsy_m[,sp_name]
  # r_pearson <- cor.test(tape_s, biopsy_s, method = "pearson") %>% broom::tidy()
  r_spearman_model <- 
    cor.test(tape_s, biopsy_s, method = "spearman") %>% 
    broom::tidy()
  
  rho <- r_spearman_model$estimate
  p <- r_spearman_model$p.value
  
  r <- list("sample" = sp_name, 
            # "r_pearson"= r_pearson, 
            "r_spearman" = rho,
            "p_value" = p)
  return(r)
})

tape_biopsy_s_res_t <- 
  tape_biopsy_s_res %>% purrr::reduce(bind_rows) %>% 
  left_join(colData(se_tape) %>% as_tibble(), 
            by = c("sample" = "subject_visit_skintype")) %>% 
  group_by(skin_type)

lm(r_spearman ~ log10(counts_sum) + skin_type, 
   data = tape_biopsy_s_res_t) %>% 
  broom::tidy()
cor.test(tape_biopsy_s_res_t$r_spearman,
         tape_biopsy_s_res_t$counts_sum)

RNAyield_cor_g <-
  tape_biopsy_s_res_t %>% 
  rename(`Tissue state` = "skin_type") %>% 
  ggscatter(y = "r_spearman", 
            x = "counts_sum", 
            color = "Tissue state",
            palette = c("LS" = "#eb2d0c", 
                        "NL" = "#eb8b9b", 
                        "HC" = "#91cf60"), 
            xscale = "log10", 
            add = "reg.line",
            add.params = list(color = "grey"),
            ylab = "Correlation coefficient",
            xlab = "Total counts") +
  scale_x_log10(labels = scales::label_log())

ggsave("../figure/F4_RNAyield_cor.jpeg",
       RNAyield_cor_g, width = 5, height = 5)
```

### Gene
```{r gene-wise correlation}
s_to_c <- intersect(colnames(expr_tape_m), colnames(expr_biopsy_m))
expr_tape_m_a <- 
  expr_tape_m[,s_to_c]
expr_biopsy_m_a <-
  expr_biopsy_m[,s_to_c]

tape_biopsy_g_res <- 
  bplapply(1:nrow(expr_tape_m_a),
  function(g_index){
  tape_g <- expr_tape_m_a[g_index,]
  biopsy_g <- expr_biopsy_m_a[g_index,]
  # r_pearson <- cor(tape_g, biopsy_g, method = "pearson")
  r_spearman <- 
    cor.test(tape_g, biopsy_g, method = "spearman") %>% broom::tidy()
  r <- list("gene" = rownames(expr_tape_m_a)[g_index], 
            # "r_pearson"= r_pearson, 
            "r_spearman" = r_spearman)
  return(r)
})

# result    -------------------
tape_biopsy_g_res_t <- 
  tibble(gene = tape_biopsy_g_res %>% map_chr(~ .x $gene),
         spearman_correlation_coefficient = tape_biopsy_g_res %>% map_dbl(~ .x $r_spearman$estimate),
         p_value = tape_biopsy_g_res %>% map_dbl(~ .x $r_spearman$p.value))

tibble(
  gene_name = rownames(expr_biopsy_m),
  tape_s = log2(expr_tape_m[, "AD_25_05_LS"]),
  biopsy_s = log2(expr_biopsy_m[, "AD_25_05_LS"])
) %>% 
ggplot(aes(tape_s, biopsy_s)) + geom_point() + 
  geom_smooth(method = "lm")

tibble(
  # gene_name = "PSME1",
  tape_s = log2(expr_tape_m_a["ALOX5AP",]),
  biopsy_s = log2(expr_biopsy_m_a["ALOX5AP",])
) %>% 
  ggplot(aes(tape_s, biopsy_s)) + geom_point()


```

## Compare with Julie's & Emma's 
```{r load data (Hu and Julies)}

# Emma
Emma_raw <- 
  openxlsx::read.xlsx("../data/Guttman-Yassky_JAMA_Dermatol_2019_Figure3.xlsx", startRow = 2) %>% as_tibble()

Emma_raw_sig <-
  Emma_raw %>% 
  mutate(across(2:4, str_detect, pattern = "a|b|c"))

Emma_raw_FC <-
  Emma_raw %>% 
  mutate(across(2:4, str_remove, "a|b|c"))

Emma_DEG_g <- 
  Emma_raw$Biomarker[Emma_raw_sig$NLvsN | Emma_raw_sig$LSvsN | Emma_raw_sig$LSvsNL]

Emma_DEG <- 
  Emma_raw_FC %>% filter(Biomarker %in% Emma_DEG_g) %>% 
  mutate(direction = ifelse(LSvsN > 0, "up", "down"))

# Dyjack ---
Dyjack <-
  openxlsx::read.xlsx("../data/Dyjack.xlsx", startRow = 2) %>% 
  as_tibble()


# Hu ----
FC_Hu <- 
  readr::read_rds("../data/FC_LSvsNL_voom.rds") %>% 
  select(feature, logFC) %>% 
  filter(feature %in% g_to_compare_FC)

# Biopsy Hu ----
FC_biopsy <-
  readr::read_csv(url("https://cdn.jsdelivr.net/gh/tuhulab/Shiny_AD_time_space@master/data/table_s2.csv"))


# Julie ---
FC_Julie <- 
  readr::read_tsv("../data/Julie/tape_LSvsNL.txt") %>% 
  select(Gene_ID, Difference)



## Venn up -----
Dyjack %>% pull(Gene) %>% clipr::write_clip()
FC_Hu %>% pull(feature) %>% clipr::write_clip()
FC_Julie %>% pull(Gene_ID) %>% 
  clipr::write_clip()


# ## Venn down -----
# Dyjack %>% filter(Log2FC < 0) %>% pull(Gene) %>% clipr::write_clip()
# FC_Hu %>% filter(logFC < 0) %>% pull(feature) %>% clipr::write_clip()
# FC_Julie %>% filter(Difference < 0) %>% pull(Gene_ID) %>% 
#   clipr::write_clip()



FC_two_studies <-
  FC_Julie %>% 
  left_join(FC_Hu, by = c("Gene_ID" = "feature")) %>% 
  mutate(Gene_ID = fct_reorder(Gene_ID, logFC ))

two_studies_g <- 
  FC_two_studies %>% 
  # slice_head(n = 8) %>% 
  rename(Sølberg = Difference,
         Hu = logFC) %>% 
  pivot_longer(-Gene_ID, names_to = "Study",
               values_to = "logFC") %>% 
  # filter(Gene_ID %in% c("MAPK12", "SPRR2G")) %>% 
  ggplot(aes(Gene_ID, logFC, fill = Study)) +
  geom_col(position = position_dodge()) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=.5))
  # facet_wrap(~Gene_ID)

ggsave("../figure/SølbergvsHu.pdf", 
       two_studies_g, width = 25, height = 5)

cor.test(FC_two_studies$Difference, FC_two_studies$logFC, method = "spearman")
```


```{r Enrichment}
## enrichr -----
library(enrichR)
enrich_Dyjack <-
  enrichr(Dyjack %>% filter(Log2FC > 0) %>% pull(Gene),
          c("GO_Biological_Process_2021", "KEGG_2021_Human")) %>% reduce(bind_rows) %>% 
  mutate(study = "Dyjack 2018")

enrich_Hu <-
    enrichr(FC_Hu %>% filter(logFC > 0) %>% pull(feature),
          c("GO_Biological_Process_2021", "KEGG_2021_Human")) %>% reduce(bind_rows) %>% 
  mutate(study = "GENAD")

enrich_Julie <-
  enrichr(FC_Julie %>% filter(Difference > 0) %>% pull(Gene_ID),
          c("GO_Biological_Process_2021", "KEGG_2021_Human")) %>% reduce(bind_rows) %>% 
  mutate(study = "Sølberg 2021")

enrichment <- 
  bind_rows(enrich_Dyjack, enrich_Hu, enrich_Julie) %>% 
  as_tibble() %>% filter(Term %in% c("keratinocyte differentiation (GO:0030216)",
                                     "inflammatory response (GO:0006954)",
                                     "cytokine-mediated signaling pathway (GO:0019221)",
                                     "regulation of T cell activation (GO:0050863)",
                                     "epidermal cell differentiation (GO:0009913)")) %>% 
  select(Term, Adjusted.P.value, Odds.Ratio, study) %>% 
  mutate(`-log10p` = - log10(Adjusted.P.value),
         theme = case_when(Term %in% c("keratinocyte differentiation (GO:0030216)",
                                       "epidermal cell differentiation (GO:0009913)") ~ "Epidermal structure",
                           TRUE ~ "Inflammation"),
         study = forcats::fct_relevel(study, c("Dyjack 2018", "Sølberg 2021", "GENAD")))

enrichment_g <-
  enrichment %>% 
    ggplot(aes(study, Term, size = `-log10p`, color = `-log10p`)) +
    geom_point() +
    facet_grid(theme~., scales = "free", space = "free") +
    scale_size_area(max_size = 8) +
    scale_color_gradient2(low = "blue", mid = "pink", high = "red") +
    scale_x_discrete(position = "top") +
    theme_bw() +
    theme(axis.title.x = element_blank(), 
          axis.title.y = element_blank(),
          axis.text.y = element_text(size = 10),
          axis.text.x = element_text(size = 9))
enrichment_g
ggsave("../data/g_enrichment.jpeg", enrichment_g, width = 7.2, height = 3.5, dpi = 600)

```


```{r}
tape_biopsy_FC_g <- 
  FC_Hu %>% filter(feature %in% g_to_compare_FC) %>% 
  left_join(FC_biopsy %>% filter(gene_name %in% g_to_compare_FC,
                     contrast %in% c("LSvsNL")) %>% 
              select(gene_name, log2FoldChange),
            by = c("feature" = "gene_name")) %>% 
  mutate(feature = fct_reorder(feature, logFC )) %>% 
  rename(Biopsy = log2FoldChange,
         Tape = logFC) %>% 
  pivot_longer(-feature, names_to = "Study",
               values_to = "logFC") %>% 
  # filter(Gene_ID %in% c("MAPK12", "SPRR2G")) %>% 
  ggplot(aes(feature, logFC, fill = Study)) +
  geom_col(position = position_dodge()) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=.5))
ggsave("../figure/BiopsyvsTape.pdf", 
       tape_biopsy_FC_g, width = 25, height = 5)

```


## Visualize

```{r}
tape_biopsy_expr <- 
  bplapply(which(colnames(expr_tape_m) %in% intersect(
  colnames(expr_tape_m), colnames(expr_biopsy_m))), 
  function(sp_index){
  sp_name <- colnames(expr_tape_m)[sp_index]
  tape_s <- expr_tape_m[,sp_name]
  biopsy_s <- expr_biopsy_m[,sp_name]
  r <- list("sample" = sp_name, 
            expr_tape = tape_s, 
            expr_biopsy = biopsy_s)
  return(r)
})

# tape_biopsy_expr %>% reduce(bind_rows) %>% 
#   filter(sample == )
```


```{r}
cor_group_g <-
  tape_biopsy_s_res_t %>% 
  ungroup %>% 
  ggboxplot(x = "skin_type", 
            y = "r_spearman", 
            add = c("jitter"), 
            fill = "skin_type",
            palette = 
              c("LS" = "#eb2d0c", 
                "NL" = "#eb8b9b", 
                "HC" = "#91cf60"),
            legend = "none", 
            xlab = "Tissue state",
            ylab = "Correlation coefficient") + 
  stat_compare_means(comparisons = list(
    c("HC", "NL"),  
    c("NL", "LS"),
    c("HC", "LS")
))  
ggsave("../figure/correlation_group.jpeg",
      cor_group_g, width = 5, height=5)
```
